requirements.txt
pandas==2.2.1
streamlit==1.33.0
python-dotenv==1.0.1
langchain-openai==0.1.3
google-cloud-bigquery==3.20.1
langchain_google_genai==1.0.2
db-dtypes==1.2.0
vertexai==1.47.0
google-generativeai==0.5.0
transformers==4.40.1
sentence-transformers==2.7.0
torch==2.1.2